requirements.txt
# Django Web Framework
Django==5.2.8

# Configuration Management
python-decouple==3.8

# Production Server
gunicorn==23.0.0

# Static Files Serving
whitenoise==6.8.2

# QR Code Generation
qrcode[pil]==8.0
Pillow==11.0.0
